refactor: dedupe groupByYear between ProjectsPage and ProjectSummaryList (#1053)
Both files had a byte-identical groupByYear() (only difference was an
explicit return-type annotation on one copy) -- the 11-line duplicate
block SonarCloud flagged between them. Extracted to lib/utils.ts as a
generic <T> version (was hard-coded to ProjectSummary[] in both, but
the logic never actually depended on the item type beyond the
getDate extractor).

Second slice of #1053's 8 tracked duplicate-block pairs; 6 remain
after this one.
