+++ frontend/src/pages/ProjectsPage.tsx
@@ -4,6 +4,7 @@ import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { useTranslation } from "react-i18next";
 import { listProjects, projectDrawdownPreviewUrl, PROJECT_TYPE_LABELS, PROJECT_STATUS_LABELS, type ProjectSummary } from "@/api/projects";
 import { AppIcons } from "@/lib/icons";
+import { groupByYear } from "@/lib/utils";
 import { previewUrl } from "@/api/drafts";
 import { AuthedImage } from "@/components/ui/AuthedImage";
 import { CreateProjectModal } from "@/components/projects/CreateProjectModal";
@@ -23,22 +24,6 @@ const STATUS_COLORS: Record<string, string> = {
 
 function fmtDate(iso: string): string {
   return new Date(iso).toLocaleDateString(undefined, { month: "short", day: "numeric" });
-}
-
-function groupByYear(
-  items: ProjectSummary[],
-  getDate: (p: ProjectSummary) => string | null,
-): Array<{ year: number; items: ProjectSummary[] }> {
-  const map = new Map<number, ProjectSummary[]>();
-  for (const item of items) {
-    const d = getDate(item);
-    const year = d ? new Date(d).getFullYear() : new Date().getFullYear();
-    if (!map.has(year)) map.set(year, []);
-    map.get(year)!.push(item);
-  }
-  return Array.from(map.entries())
-    .sort(([a], [b]) => b - a)
-    .map(([year, items]) => ({ year, items }));
 }
 
 function ProjectCard({ project, onAssign }: {
